Implémentation de la classe Carte regroupant les fonction de Decodage.cs avec ajout d'implémentation objet pour les parcelles et les cases

+++ Arabica/ArabicaLibrary/case.cs
@@ -8,12 +8,12 @@ namespace ArabicaLibrary
 {
     public class Case
     {
-        private byte x, y;
+        private int x, y;
         private char parcelle;
         private bool plantee;
         private IA proprietaire;
 
-        public Case(byte x, byte y, char parcelle)
+        public Case(int x, int y, char parcelle)
         {
             this.x = x;
             this.y = y;
@@ -22,8 +22,8 @@ namespace ArabicaLibrary
             proprietaire = null;
         }
 
-        public byte X { get => x; set => x = value; }
-        public byte Y { get => y; set => y = value; }
+        public int X { get => x; set => x = value; }
+        public int Y { get => y; set => y = value; }
         public char Parcelle { get => parcelle; set => parcelle = value; }
         public bool Plantee { get => plantee; set => plantee = value; }
         public IA Proprietaire { get => proprietaire; set => proprietaire = value; }
